+++ tests/acceptance/FoodsaverCest.php
@@ -25,6 +25,10 @@ class FoodsaverCest
 
 		$I->login($this->orga['email']);
 		$I->amOnPage('/?page=foodsaver&a=edit&id=' . $fsId);
+		$I->selectOption('Benutzerrolle', 'Foodsaver/in');
+		$I->click('Senden');
+
+		$I->amOnPage('/?page=foodsaver&a=edit&id=' . $fsId);
 		$I->selectOption('Benutzerrolle', 'Foodsharer/in');
 		$I->click('Senden');
 
@@ -35,5 +39,6 @@ class FoodsaverCest
 		$I->dontSeeInDatabase('fs_abholer', ['foodsaver_id' => $fsId]);
 		$I->dontSeeInDatabase('fs_foodsaver_has_conversation', ['foodsaver_id' => $fsId]);
 		$I->seeNumRecords(7, 'fs_quiz_session', ['foodsaver_id' => $fsId, 'quiz_id' => Role::FOODSAVER, 'status' => SessionStatus::FAILED]);
+		$I->seeInDatabase('fs_foodsaver', ['rolle' => Role::FOODSHARER, 'quiz_rolle' => Role::FOODSHARER]);
 	}
 }